refactor: move startup wiring into library
